--- packages/logging/types.ts
@@ -1,42 +1,0 @@
-/**
- * A set of logging levels
- */
-export const enum LogLevel {
-    Verbose = 0,
-    Info = 1,
-    Warning = 2,
-    Error = 3,
-    Off = 99,
-}
-
-/**
- * Interface that defines a log entry
- *
- */
-export interface ILogEntry {
-    /**
-     * The main message to be logged
-     */
-    message: string;
-    /**
-     * The level of information this message represents
-     */
-    level: LogLevel;
-    /**
-     * Any associated data that a given logging listener may choose to log or ignore
-     */
-    data?: any;
-}
-
-/**
- * Interface that defines a log listner
- *
- */
-export interface ILogListener {
-    /**
-     * Any associated data that a given logging listener may choose to log or ignore
-     *
-     * @param entry The information to be logged
-     */
-    log(entry: ILogEntry): void;
-}
